backend/apps/worker/processors/upscale.py
"""
Upscale processor for image super-resolution.
"""
import os
import requests
from PIL import Image
from typing import Dict, Any
import tempfile
from apps.core.exceptions import JobProcessingError


class UpscaleProcessor:
    """Upscale processor for image super-resolution."""
    
    def __init__(self):
        self.model_cache = {}
    
    def process(self, input_image_url: str, parameters: Dict[str, Any]) -> str:
        """Process image upscaling."""
        try:
            # Download input image
            input_path = self._download_image(input_image_url)
            
            # Get parameters
            scale_factor = parameters.get("scale_factor", 2)
            model = parameters.get("model", "esrgan")
            
            # Process upscaling
            output_path = self._process_upscale(input_path, scale_factor, model)
            
            # Upload result to S3 (placeholder - would use actual S3 upload)
            result_url = self._upload_result(output_path)
            
            # Cleanup temp files
            self._cleanup_files([input_path, output_path])
            
            return result_url
            
        except Exception as e:
            raise JobProcessingError(f"Upscaling failed: {str(e)}")
    
    def _download_image(self, image_url: str) -> str:
        """Download image from URL to temporary file."""
        response = requests.get(image_url)
        response.raise_for_status()
        
        with tempfile.NamedTemporaryFile(delete=False, suffix=".jpg") as tmp_file:
            tmp_file.write(response.content)
            return tmp_file.name
    
    def _process_upscale(self, input_path: str, scale_factor: int, model: str) -> str:
        """Process image upscaling (placeholder implementation)."""
        # In production, this would use actual super-resolution models like ESRGAN
        output_path = input_path.replace(".jpg", f"_upscale_{scale_factor}x.jpg")
        
        # Simulate upscaling processing
        image = Image.open(input_path)
        
        # Calculate new dimensions
        new_width = int(image.width * scale_factor)
        new_height = int(image.height * scale_factor)
        
        # Apply upscaling (placeholder for actual AI upscaling)
        if model == "esrgan":
            upscaled_image = self._esrgan_upscale(image, new_width, new_height)
        elif model == "real_esrgan":
            upscaled_image = self._real_esrgan_upscale(image, new_width, new_height)
        else:
            # Fallback to basic resize
            upscaled_image = image.resize((new_width, new_height), Image.Resampling.LANCZOS)
        
        upscaled_image.save(output_path, quality=95)
        return output_path
    
    def _esrgan_upscale(self, image: Image.Image, width: int, height: int) -> Image.Image:
        """ESRGAN upscaling (placeholder implementation)."""
        # In production, this would use actual ESRGAN model
        return image.resize((width, height), Image.Resampling.LANCZOS)
    
    def _real_esrgan_upscale(self, image: Image.Image, width: int, height: int) -> Image.Image:
        """Real-ESRGAN upscaling (placeholder implementation)."""
        # In production, this would use actual Real-ESRGAN model
        return image.resize((width, height), Image.Resampling.LANCZOS)
    
    def _upload_result(self, file_path: str) -> str:
        """Upload result image to S3 (placeholder implementation)."""
        # In production, this would upload to actual S3 bucket
        filename = os.path.basename(file_path)
        return f"https://oneshot-images.s3.amazonaws.com/results/{filename}"
    
    def _cleanup_files(self, file_paths: list):
        """Clean up temporary files."""
        for file_path in file_paths:
            try:
                if os.path.exists(file_path):
                    os.unlink(file_path)
            except Exception:
                pass  # Ignore cleanup errors